client-simulator/requirements.txt
typer==0.4.0
python-chess==1.7.0
